User can now only delete their own comments

+++ src/main/resources/static/js/forum.js
@@ -492,32 +492,35 @@ function generateCommentHeader(container, dto) {
         .text(dto.commenterName)
         .appendTo(row1Col1);
 
-    var row1Col2 = $("<div>")
-        .attr("class", "col-md-6")
-        .appendTo(row1);
+    var userId = $("#userId").val();
+    if (dto.commenterId === userId) {
+        var row1Col2 = $("<div>")
+            .attr("class", "col-md-6")
+            .appendTo(row1);
 
-    var row1Col1Container = $("<div>")
-        .attr("class", "d-grid gap-2 d-md-flex justify-content-md-end")
-        .appendTo(row1Col2);
+        var row1Col1Container = $("<div>")
+            .attr("class", "d-grid gap-2 d-md-flex justify-content-md-end")
+            .appendTo(row1Col2);
 
-    var deleteCommentBtn = $("<a>")
-        .attr({
-            "href": "/forum/api" + commentURI + "/" + dto.id,
-            "role": "button",
-            "class": "btn btn-danger",
-            "id": "commentDeleteBtn" + dto.id
-        })
-        .text("Delete")
-        .appendTo(row1Col1Container);
+        var deleteCommentBtn = $("<a>")
+            .attr({
+                "href": "/forum/api" + commentURI + "/" + dto.id,
+                "role": "button",
+                "class": "btn btn-danger",
+                "id": "commentDeleteBtn" + dto.id
+            })
+            .text("Delete")
+            .appendTo(row1Col1Container);
 
-    var deleteIcon = $("<i>")
-        .attr("class", "fas fa-trash")
-        .appendTo(deleteCommentBtn);
+        var deleteIcon = $("<i>")
+            .attr("class", "fas fa-trash")
+            .appendTo(deleteCommentBtn);
 
-    deleteCommentBtn.on("click", function(event) {
-        event.preventDefault();
+        deleteCommentBtn.on("click", function(event) {
+            event.preventDefault();
 
-        var deleteCommentURI = $(this).attr("href");
-        deleteComment(deleteCommentURI);
-    });
+            var deleteCommentURI = $(this).attr("href");
+            deleteComment(deleteCommentURI);
+        });
+    }
 }
